+++ e2e-tests/tests/payments_lifecycle_e2e.rs
@@ -100,7 +100,7 @@ async fn create_ar_invoice(
     pool: &PgPool,
     app_id: &str,
     customer_id: i32,
-    amount_cents: i32,
+    amount_cents: i64,
     currency: &str,
 ) -> i32 {
     sqlx::query_scalar::<_, i32>(
