+++ e2e-tests/tests/reconciliation_matching_e2e.rs
@@ -48,7 +48,7 @@ async fn create_invoice(
     pool: &sqlx::PgPool,
     tenant_id: &str,
     customer_id: i32,
-    amount_cents: i32,
+    amount_cents: i64,
     currency: &str,
 ) -> i32 {
     sqlx::query_scalar::<_, i32>(
@@ -76,7 +76,7 @@ async fn create_invoice_with_ref(
     pool: &sqlx::PgPool,
     tenant_id: &str,
     customer_id: i32,
-    amount_cents: i32,
+    amount_cents: i64,
     currency: &str,
     tilled_invoice_id: &str,
 ) -> i32 {
@@ -105,7 +105,7 @@ async fn create_charge(
     pool: &sqlx::PgPool,
     tenant_id: &str,
     customer_id: i32,
-    amount_cents: i32,
+    amount_cents: i64,
     currency: &str,
     reference_id: Option<&str>,
 ) -> i32 {
